pyproject.toml
[project]
name = "spase-cache"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fitter",
    "hydra-core>=1.3.2",
    "matplotlib",
    "numpy>=2.2.6",
    "polars",
    "torch",
    "transformers>=5.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
